link state
